enhancement (<OptimizelyFeature>): Convert to functional component that uses useFeature hook (#32)
Summary
Updates <OptimizelyFeature> to be a functional component that

Also adds support for userId and userAttribute overrides.

Addresses #30

Release
Is this a breaking change? In theory the logic should behave the same, as #28 was designed to implement the same logic as <OptimizelyFeature>.

Timing wise, when isServerSide is false, the state setting of the correct values of the feature/variables in useFeature happens in a second chained thennable, rather than the first.

I dont think this should be considered a breaking change (and thus warrant a major release), but I wanted to point it out.