fix(hooks): Always recompute decision after resolution of ready promise (#91)
Summary:

Fix the following bug: when client is initialized with datafile, and no SDK key, OptimizelyExperiment and OptimizelyFeature components don't properly render their decisions.

The issue is in the implementation in hooks and the interaction with the client's ready state (via isReady/onReady). When initialized with datafile, the client begins in a non-ready state, but immediately becomes ready on the next tick. But the hook does not recompute the decision when this occurs - it's only recomputing the decision when a CONFIG_UPDATE is fired. CONFIG_UPDATE is never fired when the instance was initialized with datafile only.

To fix, add an effect in the useFeature/useExperiment hooks that recomputes the decision after the client becomes ready.

Test plan:

- Manual testing
- Updated unit tests