chore: Adding E2E tests (#151)
## Summary
React SDK contains unit tests but its not possible to cover many scenarios using those. We have started working on an E2E test suite in a private repo [here](https://github.com/optimizely/react-sdk-e2e-tests). This PR enables this repo to trigger E2E tests on PRs. E2E test suite is not complete yet, so we are not making the check mandatory. Once we have something significant in that repo, we will make this check mandatory.

Since we were working in the travis config file, we cleaned it up a bit and removed the unnecessary markdown lint which has already been removed from other repos.

## Test Plan
Tested the triggering mechanism thoroughly and it appears to be working fine.